+++ getRecentTracks.php
@@ -1,8 +1,8 @@
 <?
   require('db.php');
   $data = json_decode(file_get_contents('php://input'));
-  $page = mysqli_real_escape_string($link, $data->{'page'});
-  $overrideMaxResults = mysqli_real_escape_string($link, $data->{'maxResultsPerPage'});
+  $page = (int)mysqli_real_escape_string($link, $data->{'page'});
+  $overrideMaxResults = (int)mysqli_real_escape_string($link, $data->{'maxResultsPerPage'});
   if($overrideMaxResults) $maxResultsPerPage = $overrideMaxResults;
 
   $start = $maxResultsPerPage * $page;
